# Implement Decoder backward self-overlapping