### PR types
Performance optimization

### PR changes
Models

### Description
实现 Decoder backward 的<b>【自Overlap】</b>机制，即：
1. ATTN(W)+POST(B) 与 COMBINE(B) 进行 overlap
2. MLP(W) 与 DISPATCH(B) 进行 overlap

下图展示了一个 backward chunk 的原版和自Overlap的编排对比：
<img width="1835" height="460" alt="图片 1" src="https://github.com/user-attachments/assets/99c2ca2c-dd43-4463-b2d8-b91b65634b54" />

实际 timeline 看起来挺好，基本能 overlap 住：
<img width="1277" height="308" alt="图片 2" src="https://github.com/user-attachments/assets/041073e7-b666-49b5-8765-45642098549f" />

<b>性能：</b>pp4ep8，16 mini-batch
单 chunk 用时 112ms -> 96ms，提速 14%
单 batch 用时 4.834s -> 4.791s，提速 1%

该 PR 未实现 send/recv overlap，传入的 pp_stream 始终为 None，但保留了传入的接口；当然这个的提升估计会非常小，就像之前做 forward 阶段的 send/recv overlap 一样

该 PR 需要 Paddle 更新到 https://github.com/PaddlePaddle/Paddle/pull/74891 才能生效，但不更新也不会错，会跑原来无 overlap 的分支，性能不变